pyproject.toml
[tool.poetry]
name = "vivek451-quiz"
version = "0.1.0"
description = ""
authors = ["vivek451"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2.7"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

